requirements.txt
pyyaml
flask
matplotlib
numpy
pandas
scipy
scikit-learn